+++ stable/uys.h
@@ -434,12 +434,13 @@ namespace uys {
 		return m_Undefined;	// Pour viter un 'warning'.
 	}
 
-	E_ENUM( behavior )
+	qENUM( Behavior )
 	{
 		bVolatile,
-			bPersistent,
-			b_amount,
-			b_Undefined
+		bPersistent,
+		b_amount,
+		b_Undefined,
+		b_Default = bPersistent
 	};
 }
 
@@ -457,10 +458,18 @@ namespace uys {
 			Filename.reset( P );
 		}
 		E_CDTOR( rHF );
+		rHF(
+      const fnm::rName &Path,
+      const fnm::rName &Basename = fnm::Empty)
+		{
+		  Init(Path, Basename);
+		}
 		void Init(
 			const fnm::name___ &Path,
-			const fnm::name___ &Basename )	// Peut tre vide ('NULL') si 'Path' contient dj le nom de fichier.
+			const fnm::name___ &Basename = NULL)	// Peut tre vide si 'Path' contient dj le nom de fichier.
 		{
+		  reset();
+
 			Filename.Init();
 			fnm::BuildPath( Path, Basename, ".q37", Filename );
 		}
@@ -479,66 +488,79 @@ namespace uys {
 
 	qXENUM( State, s );
 
-	// Files hook.
-	class rFH
-	: public flsq::file_storage_driver___
+	// Generic files hook.
+	template <typename fsd> class rFH_
+	: public fsd
 	{
 	public:
 		void reset( bso::sBool P = true )
 		{
 			if ( P ) {
-				if ( IsInitialized() ) {
+				if ( fsd::IsInitialized() ) {
 					if ( !State().Boolean() )
-						if ( IsPersistent() && ( Mode() == mReadWrite ) )
-							CreateFile();
+						if ( fsd::IsPersistent() && ( Mode() == mReadWrite ) )
+							fsd::CreateFile();
 				}
 			}
 
-			flsq::file_storage_driver___::reset( P );
+			fsd::reset( P );
 		}
-		qCVDTOR( rFH );
+		qCVDTOR( rFH_ );
 		eState Init(
 			const rHF &Filenames,
 			mode__ Mode,
-			behavior__ Behavior,
-			flsq::id__ ID )
+			eBehavior Behavior = b_Default,
+			flsq::rId Id = flsq::Undefined)
 		{
-			flsq::file_storage_driver___::Init( ID, Filenames.Filename, Convert_( Mode ), flsq::cFirstUse );
+			fsd::Init(Id, Filenames.Filename, Convert_( Mode ), flsq::cFirstUse);
 
 			switch ( Behavior ) {
 			case bVolatile:
 				break;
 			case bPersistent:
-				Persistent();
+				fsd::Persistent();
 				break;
 			default:
 				qRFwk();
 				break;
 			}
 
-			if (FileExists() )
+			if ( fsd::FileExists() )
 				return sExists;
 			else
 				return sAbsent;
 		}
+		eState Init(
+			const fnm::rName &Name,
+			mode__ Mode,
+			eBehavior Behavior = b_Default,
+			flsq::rId Id = flsq::Undefined)
+		{
+		  return Init(rHF(Name), Mode, Behavior, Id);
+		}
 		eState State( void ) const
 		{
-			return ( FileExists() ? sExists : sAbsent );
+			return fsd::FileExists() ? sExists : sAbsent;
 		}
-
 		void DropFiles( void )
 		{
-			Drop();
+			fsd::Drop();
 		}
 		mode__ SetMode( mode__ Mode )
 		{
-			return Convert_( flsq::file_storage_driver___::Mode( Convert_( Mode ) ) );
+			return Convert_( fsd::Mode( Convert_( Mode ) ) );
 		}
 		mode__ Mode( void ) const
 		{
-			return Convert_( flsq::file_storage_driver___::Mode() );
+			return Convert_( fsd::Mode() );
 		}
 	};
+
+	// Regular file hook.
+	typedef rFH_<flsq::rFileDriver> rFH;
+
+	// File hook with offset.
+  template <int offset> qTCLONE(rFH_<flsq::rFileOffsetDriver<offset>>, rFOH);
 
 	class untyped_storage
 	: public untyped_storage_